+++ Firmament/src/utopia/firmament/factory/VariableFramedFactory.scala
@@ -1,6 +1,7 @@
 package utopia.firmament.factory
 
 import utopia.firmament.model.stack.{StackInsets, StackInsetsConvertible}
+import utopia.flow.util.Mutate
 import utopia.flow.view.immutable.eventful.Fixed
 import utopia.flow.view.template.eventful.Changing
 
@@ -17,7 +18,6 @@ trait VariableFramedFactory[+A] extends FramedFactory[A]
 	  * @return A pointer that contains the applied insets
 	  */
 	def insetsPointer: Changing[StackInsets]
-	
 	/**
 	  * @param p New insets pointer to apply
 	  * @return Copy of this factory with the specified variable insets applied
@@ -28,16 +28,20 @@ trait VariableFramedFactory[+A] extends FramedFactory[A]
 	// IMPLEMENTED  ---------------------
 	
 	override def withInsets(insets: StackInsetsConvertible): A = withInsetsPointer(Fixed(insets.toInsets))
-	
-	override def mapInsets(f: StackInsets => StackInsetsConvertible): A =
-		withInsetsPointer(insetsPointer.map { f(_).toInsets })
+	override def mapInsets(f: StackInsets => StackInsetsConvertible): A = mapInsetsPointer { _.map { f(_).toInsets } }
 		
 	
 	// OTHER    -------------------------
 	
 	/**
+	  * @param f A mapping function applied to the insets-pointer
+	  * @return Copy of this factory using the mapped/modified pointer
+	  */
+	def mapInsetsPointer(f: Mutate[Changing[StackInsets]]) = withInsetsPointer(f(insetsPointer))
+	
+	/**
 	  * @param f A mapping function for insets, yielding a variable result
 	  * @return Copy of this factory applying the specified mapping function to insets
 	  */
-	def flatMapInsets(f: StackInsets => Changing[StackInsets]) = withInsetsPointer(insetsPointer.flatMap(f))
+	def flatMapInsets(f: StackInsets => Changing[StackInsets]) = mapInsetsPointer { _.flatMap(f) }
 }